Revert "Diego kpmg"
